requirements.txt
kubernetes==1.0.0b1 
jinja2~=2.9.5
pyyaml~=3.12
click~=6.7
